update student by mentor

+++ controllers/batches.js
@@ -5,17 +5,38 @@ module.exports = {
   /* Mentor Routes */
   create: async (req, res, next) => {
     try {
-      const batch = await Batch.create(req.body);
-      return res.json({ success: true, batch });
+      if(req.isMentor) {
+        const batch = await Batch.create(req.body);
+        return res.json({ success: true, batch });
+      } else {
+        return res.json({ msg: 'Not Authorized' });
+      }
     } catch (err) {
+      next(err);
+    }
+  },
+
+  update: async (req, res, next) => {
+    try {
+      if(req.isMentor) {
+        const batch = await Batch.findByIdAndUpdate(req.params.id, req.body, {new: true});
+        return res.json({success: true, batch});
+      } else {
+        return res.json({ msg: 'Not Authorized' });
+      }
+    } catch(err) {
       next(err);
     }
   },
 
   delete: async (req, res, next) => {
     try {
-      const batch = await Batch.findByIdAndDelete(req.params.id);
-      return res.json({ success: true, batch });
+      if(req.isMentor) {
+        const batch = await Batch.findByIdAndDelete(req.params.id);
+        return res.json({ success: true, batch });
+      } else {
+        return res.json({ msg: 'Not Authorized' });
+      }
     } catch (err) {
       next(err);
     }
